Prezzi long altrimenti le moltiplicazioni facevano casino e mappa dei barcodes nel carrello

+++ PromoCalculator/BaseTypes.h
@@ -11,7 +11,7 @@
 
 struct Totals {
     long itemsNumber ;
-    double totalAmount ;
+    long totalAmount ;
 };
 
 struct CartRow {
